Extracted the CE Service in a sepparate class
