+++ pom.xml
@@ -28,7 +28,7 @@
 		<jacoco.version>0.8.8</jacoco.version>
 		<jsonwebtoken.version>0.11.5</jsonwebtoken.version>
 		<jackson.databind.version>2.13.2</jackson.databind.version>
-		<log4j2.version>2.17.1</log4j2.version>
+		<log4j2.version>2.17.2</log4j2.version>
 	</properties>
 
 	<dependencyManagement>
@@ -265,7 +265,7 @@
 		<dependency>
 			<groupId>io.boomerang</groupId>
 			<artifactId>lib-eventing</artifactId>
-			<version>0.3.1</version>
+			<version>0.4.3</version>
 		</dependency>
 		<dependency>
 			<groupId>io.boomerang</groupId>
