+++ src/main/java/io/boomerang/service/CloudEventsService.java
@@ -1,0 +1,8 @@
+package io.boomerang.service;
+
+import java.util.InvalidPropertiesFormatException;
+import io.cloudevents.CloudEvent;
+
+public interface CloudEventsService {
+  void processCloudEventRequest(CloudEvent cloudEvent) throws InvalidPropertiesFormatException;
+}
